Disable/enable next button when changing interaction mode
The next button is enabled or disabled based on whether the player is using the correct interaction mode.
